+++ src/index.js
@@ -2,6 +2,7 @@ import Button from './Button.svelte';
 import ToggleButton from './ToggleButton.svelte';
 import ButtonGroup from './ButtonGroup.svelte';
 import Checkbox from './Checkbox.svelte';
+import Datefield from './Datefield.svelte';
 import Datepicker from './Datepicker.svelte';
 import Dialog from './Dialog.svelte';
 import Icon from './Icon.svelte';
@@ -17,6 +18,7 @@ export {
 	Button,
 	ButtonGroup,
 	Checkbox,
+	Datefield,
 	Datepicker,
 	Dialog,
 	Icon,