+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Datefield
+- Popover
+	- dispatch 'close' custom event
 - Datepicker
 
 ## 0.1.11